+++ src/company/Data/Station.java
@@ -13,7 +13,7 @@ public class Station {
     private int id;
     private String name;
     private int nahieh;
-    private double capacity;
+
     private ArrayList<Long> stationWagon;
     private HashMap<Integer, Capacity> stationCapacity;
     private HashMap<TimeHashMapKey, WagonTypeTime> timeHashMap;
@@ -33,6 +33,7 @@ public class Station {
         this.name = name;
         this.nahieh = nahieh;
         this.stationWagon= new ArrayList<>();
+        this.stationCapacity= new HashMap<>();
     }
 
     public int getNahieh() {
@@ -65,14 +66,6 @@ public class Station {
 
     public void setStationWagon(ArrayList<Long> stationWagon) {
         this.stationWagon = stationWagon;
-    }
-
-    public double getCapacity() {
-        return capacity;
-    }
-
-    public void setCapacity(double capacity) {
-        this.capacity = capacity;
     }
 
     public HashMap<TimeHashMapKey, WagonTypeTime> getTimeHashMap() {
@@ -153,9 +146,18 @@ public class Station {
         }
     }
 
-    private class Capacity {
-        int loadingCap;
-        int unloadingCap;
-        int oncomingWagons;
+    public HashMap<Integer, Capacity> getStationCapacity() {
+        return stationCapacity;
+    }
+
+    public void setStationCapacity(HashMap<Integer, Capacity> stationCapacity) {
+        this.stationCapacity = stationCapacity;
+    }
+
+    public class Capacity {
+        public int loadingCap;
+        public int unloadingCap;
+        public int comingLoadWagons;
+        public int comingEmptyWagons;
     }
 }
